+++ tests/testthat/test_checkValue.R
@@ -4,18 +4,21 @@
 load(file = "helper_data.rda")
 
 df3 <- df2
-df3$dat[1, 1] <- 8
+df3$dat[1, 1:2] <- 8
+
+test_that("Input validation", {
+  expect_error(checkValue(df1, value = 3:4),
+               "'value' needs to be of length 1.")
+})
 
 test_that("Value checks raise no false alarms", {
-  expect_silent(checkValue(df1, value = 4))
-  expect_silent(checkValue(df2, value = -1))
+  expect_equal(checkValue(df1, value = 4), integer(0))
+  expect_equal(checkValue(df2, value = -1), integer(0))
 })
 
 test_that("Value occurences reported", {
-  expect_message(checkValue(df1, value = 1),
-                 "The following variables have occurences of 'value' in the GADSdat:
-ID1")
-  expect_message(checkValue(df3, value = 8),
-                 "The following variables have occurences of 'value' in the GADSdat:
-ID1, V2")
+  expect_equal(checkValue(df1, value = 1),
+                 c(ID1 = 1L))
+  expect_equal(checkValue(df3, value = 8),
+                 c(ID1 = 1L, V2 = 2L))
 })