+++ NEWS.md
@@ -1,6 +1,7 @@
 # eatGADS 0.15.2.9000
 
 * bug fix `collapseColumns()` (now supports all column names in look up table)
+* `checkValue()` now returns occurence count as vector instead of giving a message
 
 # eatGADS 0.15.2